+++ pkgs/shelf_router/lib/src/router.dart
@@ -132,10 +132,6 @@ class Router {
     });
   }
 
-  /// Get a [Handler] that will route incoming requests to registered handlers.
-  @Deprecated('The Router class is a Handler on its own')
-  Handler get handler => call;
-
   /// Route incoming requests to registered handlers.
   ///
   /// This method allows a Router instance to be a [Handler].